Fix blas1 and blas2 dot functions
blas1_t and blas_b compile, further tests should be
done to test the mixed type dispatching

+++ inc/dg/backend/exblas/exdot_omp.h
@@ -133,7 +133,7 @@ void ExDOTFPE(int N, RandomAccessIterator1 a, RandomAccessIterator2 b, int64_t* 
         int r = ((((tid+1) * int64_t(N)) / tnum) ) - 1;
         for(int i = l; i <= r; i++) {
             double r1;
-            double x = TwoProductFMA(a[i],b[i],r1);
+            double x = TwoProductFMA((double)a[i],(double)b[i],r1);
             cache.Accumulate(x);
             cache.Accumulate(r1);
         }
